OpenGL/Various: Added asserts for using the wrong context in the FBO, More use of string_view in many places, various warning fixes. Volumeraycater fixes for halding errors in the pool job. Removed templated canvasqt since we only use one backend.

+++ modules/postprocessing/include/modules/postprocessing/processors/depthoffield.h
@@ -104,8 +104,8 @@ private:
     MeshInport trackingInport_;
     ImageOutport outport_;
 
-    FloatProperty aperture_;
-    FloatProperty focusDepth_;
+    DoubleProperty aperture_;
+    DoubleProperty focusDepth_;
     BoolProperty manualFocus_;
     BoolProperty approximate_;
     IntSizeTProperty viewCountExact_;
